account history fix and rename ecency folder
