+++ src/components/postCard/container/postCardContainer.js
@@ -6,7 +6,7 @@ import get from 'lodash/get';
 // Services
 import { act } from 'react-test-renderer';
 import { getPost, getActiveVotes } from '../../../providers/hive/dhive';
-import { getPostReblogs } from '../../../providers/esteem/esteem';
+import { getPostReblogs } from '../../../providers/ecency/ecency';
 
 import { parseActiveVotes } from '../../../utils/postParser';
