+++ src/components/searchModal/container/searchModalContainer.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import get from 'lodash/get';
 
 // Services and Actions
-import { search } from '../../../providers/esteem/esteem';
+import { search } from '../../../providers/ecency/ecency';
 import { lookupAccounts, getTrendingTags, getPurePost } from '../../../providers/hive/dhive';
 
 // Constants
