+++ src/containers/profileEditContainer.js
@@ -6,7 +6,7 @@ import ImagePicker from 'react-native-image-crop-picker';
 import get from 'lodash/get';
 import { withNavigation } from 'react-navigation';
 
-import { uploadImage } from '../providers/esteem/esteem';
+import { uploadImage } from '../providers/ecency/ecency';
 
 import { profileUpdate, signImage } from '../providers/hive/dhive';
 import { updateCurrentAccount } from '../redux/actions/accountAction';
