+++ src/screens/post/container/postContainer.js
@@ -6,7 +6,7 @@ import get from 'lodash/get';
 // Services and Actions
 import Matomo from 'react-native-matomo-sdk';
 import { getPost } from '../../../providers/hive/dhive';
-// import { matomo } from '../../../providers/esteem/analytics';
+// import { matomo } from '../../../providers/ecency/analytics';
 
 // Component
 import PostScreen from '../screen/postScreen';
