+++ src/screens/register/registerContainer.js
@@ -6,7 +6,7 @@ import { withNavigation } from 'react-navigation';
 import get from 'lodash/get';
 
 import { lookupAccounts } from '../../providers/hive/dhive';
-import { register } from '../../providers/esteem/esteem';
+import { register } from '../../providers/ecency/ecency';
 import ROUTES from '../../constants/routeNames';
 
 const RegisterContainer = ({ children, navigation }) => {
