+++ src/screens/searchResult/container/otherResultContainer.js
@@ -6,7 +6,7 @@ import { connect } from 'react-redux';
 import ROUTES from '../../../constants/routeNames';
 
 import { lookupAccounts, getTrendingTags } from '../../../providers/hive/dhive';
-import { getLeaderboard } from '../../../providers/esteem/esteem';
+import { getLeaderboard } from '../../../providers/ecency/ecency';
 
 const OtherResultContainer = (props) => {
   const [users, setUsers] = useState([]);
